move printing helpers to `__init.py__`
