+++ tests/sentry/grouping/test_variants.py
@@ -1,62 +1,20 @@
 from __future__ import annotations
 
 from os import path
-from typing import Any
 
-import orjson
 import pytest
 
 from sentry.eventstore.models import Event
-from sentry.grouping.component import GroupingComponent
 from sentry.grouping.strategies.configurations import CONFIGURATIONS
-from sentry.grouping.variants import BaseVariant
 from sentry.models.project import Project
 from sentry.projectoptions.defaults import DEFAULT_GROUPING_CONFIG
 from sentry.testutils.pytest.fixtures import InstaSnapshotter, django_db_all
-from tests.sentry.grouping import GROUPING_INPUTS_DIR, GroupingInput, with_grouping_inputs
-
-
-def to_json(value: Any) -> str:
-    return orjson.dumps(value, option=orjson.OPT_SORT_KEYS).decode()
-
-
-def dump_variant(
-    variant: BaseVariant, lines: list[str] | None = None, indent: int = 0
-) -> list[str]:
-    if lines is None:
-        lines = []
-
-    def _dump_component(component: GroupingComponent, indent: int) -> None:
-        if not component.hint and not component.values:
-            return
-        lines.append(
-            "%s%s%s%s"
-            % (
-                "  " * indent,
-                component.id,
-                component.contributes and "*" or "",
-                component.hint and " (%s)" % component.hint or "",
-            )
-        )
-        for value in component.values:
-            if isinstance(value, GroupingComponent):
-                _dump_component(value, indent + 1)
-            else:
-                lines.append("{}{}".format("  " * (indent + 1), to_json(value)))
-
-    lines.append("{}hash: {}".format("  " * indent, to_json(variant.get_hash())))
-
-    for key, value in sorted(variant.__dict__.items()):
-        if isinstance(value, GroupingComponent):
-            lines.append("{}{}:".format("  " * indent, key))
-            _dump_component(value, indent + 1)
-        elif key == "config":
-            # We do not want to dump the config
-            continue
-        else:
-            lines.append("{}{}: {}".format("  " * indent, key, to_json(value)))
-
-    return lines
+from tests.sentry.grouping import (
+    GROUPING_INPUTS_DIR,
+    GroupingInput,
+    dump_variant,
+    with_grouping_inputs,
+)
 
 
 @with_grouping_inputs("grouping_input", GROUPING_INPUTS_DIR)
